update smobiler4.2 smoone新增ActionButton、Toolbar 消息、列表侧滑删除、登录界面记住密码更改为圆形样式

+++ SmoONE.UI/RB/frmRTypeTempCreate.cs
@@ -136,17 +136,23 @@ namespace SmoONE.UI.RB
         {
             try
             {
-                ReturnInfo r = AutofacConfig.rBService.DeleteRB_Type_Template(ID);
-                if (r.IsSuccess == true)
+                MessageBox.Show("是否确定删除消费模板？", MessageBoxButtons.YesNo, (Object s1, MessageBoxHandlerArgs args1) =>
                 {
-                    this.ShowResult = ShowResult.Yes;
-                    this.Close();
-                    Toast("删除消费模板成功");
-                }
-                else
-                {
-                    throw new Exception(r.ErrorInfo);
-                }
+                    if (args1.Result == Smobiler.Core.Controls.ShowResult.Yes)
+                    {
+                        ReturnInfo r = AutofacConfig.rBService.DeleteRB_Type_Template(ID);
+                        if (r.IsSuccess == true)
+                        {
+                            this.ShowResult = ShowResult.Yes;
+                            this.Close();
+                            Toast("删除消费模板成功");
+                        }
+                        else
+                        {
+                            throw new Exception(r.ErrorInfo);
+                        }
+                    }
+                });
             }
             catch (Exception ex)
             {
